+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/rewards/GetSyncCommitteeRewardsTest.java
@@ -18,7 +18,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_BAD_REQUEST;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_FOUND;
-import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_IMPLEMENTED;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
 import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.getResponseStringFromMetadata;
 import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.verifyMetadataErrorResponse;
@@ -26,10 +25,19 @@ import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.verifyMe
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.io.Resources;
 import java.io.IOException;
+import java.util.List;
+import java.util.Set;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.api.migrated.SyncCommitteeRewardData;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerWithChainDataProviderTest;
+import tech.pegasys.teku.spec.SpecMilestone;
+import tech.pegasys.teku.spec.TestSpecFactory;
+import tech.pegasys.teku.spec.datastructures.blocks.SignedBlockAndState;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.SyncAggregate;
+import tech.pegasys.teku.spec.generator.ChainBuilder;
+import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 public class GetSyncCommitteeRewardsTest
     extends AbstractMigratedBeaconHandlerWithChainDataProviderTest {
@@ -37,7 +45,44 @@ public class GetSyncCommitteeRewardsTest
 
   @BeforeEach
   void setup() {
-    setHandler(new GetSyncCommitteeRewards());
+    spec = TestSpecFactory.createMinimalAltair();
+    dataStructureUtil = new DataStructureUtil(spec);
+    initialise(SpecMilestone.ALTAIR);
+    genesis();
+    final SyncAggregate syncAggregate =
+        dataStructureUtil.randomSyncAggregate(0, 3, 4, 7, 8, 9, 10, 16, 17, 20, 23, 25, 26, 29, 30);
+    final ChainBuilder.BlockOptions blockOptions =
+        ChainBuilder.BlockOptions.create().setSyncAggregate(syncAggregate);
+    SignedBlockAndState blockAndState = chainBuilder.generateBlockAtSlot(1, blockOptions);
+    chainUpdater.saveBlock(blockAndState);
+    chainUpdater.updateBestBlock(blockAndState);
+    setHandler(new GetSyncCommitteeRewards(chainDataProvider));
+    request.setPathParameter("block_id", "head");
+  }
+
+  @Test
+  void shouldReturnSyncCommitteeRewardsInformation_emptyValidators() throws Exception {
+    request.setRequestBody(List.of());
+    handler.handleRequest(request);
+
+    final SyncCommitteeRewardData output =
+        chainDataProvider.getSyncCommitteeRewardsFromBlockId("head", Set.of()).get().orElseThrow();
+    Assertions.assertThat(request.getResponseCode()).isEqualTo(SC_OK);
+    assertThat(request.getResponseBody()).isEqualTo(output);
+  }
+
+  @Test
+  void shouldReturnSyncCommitteeRewardsInformation_specifiedValidators() throws Exception {
+    request.setRequestBody(List.of("0", "3", "9", "10"));
+    handler.handleRequest(request);
+
+    final SyncCommitteeRewardData output =
+        chainDataProvider
+            .getSyncCommitteeRewardsFromBlockId("head", Set.of("0", "3", "9", "10"))
+            .get()
+            .orElseThrow();
+    Assertions.assertThat(request.getResponseCode()).isEqualTo(SC_OK);
+    assertThat(request.getResponseBody()).isEqualTo(output);
   }
 
   @Test
@@ -53,11 +98,6 @@ public class GetSyncCommitteeRewardsTest
   @Test
   void metadata_shouldHandle500() throws JsonProcessingException {
     verifyMetadataErrorResponse(handler, SC_INTERNAL_SERVER_ERROR);
-  }
-
-  @Test
-  void metadata_shouldHandle501() throws JsonProcessingException {
-    verifyMetadataErrorResponse(handler, SC_NOT_IMPLEMENTED);
   }
 
   @Test