+++ data/provider/src/main/java/tech/pegasys/teku/api/ChainDataProvider.java
@@ -24,6 +24,7 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
 import java.io.ByteArrayInputStream;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -50,7 +51,6 @@ import tech.pegasys.teku.api.response.v1.beacon.GenesisData;
 import tech.pegasys.teku.api.response.v1.beacon.ValidatorStatus;
 import tech.pegasys.teku.api.schema.BeaconState;
 import tech.pegasys.teku.api.schema.Fork;
-import tech.pegasys.teku.api.schema.Version;
 import tech.pegasys.teku.api.stateselector.StateSelectorFactory;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
@@ -62,6 +62,7 @@ import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.SyncAggregate;
+import tech.pegasys.teku.spec.datastructures.execution.versions.capella.Withdrawal;
 import tech.pegasys.teku.spec.datastructures.forkchoice.ProtoNodeData;
 import tech.pegasys.teku.spec.datastructures.forkchoice.ReadOnlyForkChoiceStrategy;
 import tech.pegasys.teku.spec.datastructures.lightclient.LightClientBootstrap;
@@ -73,6 +74,8 @@ import tech.pegasys.teku.spec.datastructures.state.CommitteeAssignment;
 import tech.pegasys.teku.spec.datastructures.state.SyncCommittee;
 import tech.pegasys.teku.spec.datastructures.type.SszPublicKey;
 import tech.pegasys.teku.spec.logic.common.statetransition.epoch.status.ValidatorStatuses;
+import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.EpochProcessingException;
+import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.SlotProcessingException;
 import tech.pegasys.teku.storage.client.ChainDataUnavailableException;
 import tech.pegasys.teku.storage.client.CombinedChainDataClient;
 import tech.pegasys.teku.storage.client.RecentChainData;
@@ -609,6 +612,8 @@ public class ChainDataProvider {
           .collect(Collectors.<Integer, Integer, Integer>toMap(Function.identity(), result::get));
     }
 
+    checkValidatorsList(committeeKeys, validators);
+
     final Map<Integer, Integer> output = new HashMap<>();
     for (int i = 0; i < committeeKeys.size(); i++) {
       final BLSPublicKey key = committeeKeys.get(i);
@@ -621,6 +626,30 @@ public class ChainDataProvider {
     }
 
     return output;
+  }
+
+  private void checkValidatorsList(List<BLSPublicKey> committeeKeys, Set<String> validators) {
+    final Set<BLSPublicKey> keysSet = new HashSet<>(committeeKeys);
+    for (String v : validators) {
+      final String errorMessage =
+          String.format(
+              "'%s' is not a valid hex encoded public key or validator index in the committee", v);
+
+      if (v.startsWith("0x")) {
+        if (!keysSet.contains(BLSPublicKey.fromHexString(v))) {
+          throw new BadRequestException(errorMessage);
+        }
+      } else {
+        try {
+          final int index = Integer.parseInt(v);
+          if (index < 0 || index >= committeeKeys.size()) {
+            throw new BadRequestException(errorMessage);
+          }
+        } catch (NumberFormatException e) {
+          throw new BadRequestException(errorMessage);
+        }
+      }
+    }
   }
 
   @VisibleForTesting
@@ -650,8 +679,63 @@ public class ChainDataProvider {
     return spec.atSlot(slot).getMilestone();
   }
 
-  public Version getVersionAtSlot(final UInt64 slot) {
-    return Version.fromMilestone(spec.atSlot(slot).getMilestone());
+  public SafeFuture<Optional<ObjectAndMetaData<List<Withdrawal>>>> getExpectedWithdrawals(
+      String stateParameter, Optional<UInt64> optionalProposalSlot) {
+    return defaultStateSelectorFactory
+        .defaultStateSelector(stateParameter)
+        .getState()
+        .thenApply(
+            maybeStateData ->
+                maybeStateData.map(
+                    stateAndMetaData -> {
+                      List<Withdrawal> withdrawals =
+                          getExpectedWithdrawalsFromState(
+                              stateAndMetaData.getData(), optionalProposalSlot);
+                      return new ObjectAndMetaData<>(
+                          withdrawals,
+                          stateAndMetaData.getMilestone(),
+                          stateAndMetaData.isExecutionOptimistic(),
+                          stateAndMetaData.isCanonical(),
+                          stateAndMetaData.isFinalized());
+                    }));
+  }
+
+  List<Withdrawal> getExpectedWithdrawalsFromState(
+      tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState data,
+      Optional<UInt64> optionalProposalSlot) {
+    final UInt64 proposalSlot = optionalProposalSlot.orElse(data.getSlot().increment());
+    // Apply some sanity checks prior to computing pre-state
+    if (!spec.atSlot(proposalSlot).getMilestone().isGreaterThanOrEqualTo(SpecMilestone.CAPELLA)) {
+      throw new BadRequestException(
+          "Requested withdrawals for a pre-capella slot: " + proposalSlot);
+    } else if (proposalSlot.isLessThanOrEqualTo(data.getSlot())) {
+      // expectedWithdrawals is computing a forward-looking set of withdrawals. if looking for
+      // withdrawals from a state, this is the incorrect endpoint.
+      throw new BadRequestException(
+          "Requested withdrawals for a historic slot, state slot: "
+              + data.getSlot()
+              + " vs. proposal slot: "
+              + proposalSlot);
+    } else if (proposalSlot.minusMinZero(128).isGreaterThan(data.getSlot())) {
+      // 128 slots is 4 epochs, which would suggest 4 epochs of empty blocks to process
+      throw new BadRequestException(
+          "Requested withdrawals for a proposal slot: "
+              + proposalSlot
+              + " that is more than 128 slots ahead of the supplied state slot: "
+              + data.getSlot());
+    }
+    try {
+      // need to get preState
+      final tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState preState =
+          spec.processSlots(data, proposalSlot);
+      return spec.atSlot(proposalSlot)
+          .getBlockProcessor()
+          .getExpectedWithdrawals(preState)
+          .orElse(List.of());
+    } catch (SlotProcessingException | EpochProcessingException e) {
+      LOG.debug("Failed to get expected withdrawals for slot {}", proposalSlot, e);
+    }
+    return List.of();
   }
 
   public SafeFuture<Optional<Bytes32>> getFinalizedBlockRoot(final UInt64 slot) {