+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigDeneb.java
@@ -13,6 +13,8 @@
 
 package tech.pegasys.teku.spec.config;
 
+import static tech.pegasys.teku.spec.config.Constants.MAX_REQUEST_BLOCKS_DENEB;
+
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
@@ -40,6 +42,10 @@ public interface SpecConfigDeneb extends SpecConfigCapella {
   int getFieldElementsPerBlob();
 
   int getMaxBlobsPerBlock();
+
+  default UInt64 getMaxRequestBlobSidecars() {
+    return MAX_REQUEST_BLOCKS_DENEB.times(getMaxBlobsPerBlock());
+  }
 
   Optional<String> getTrustedSetupPath();