+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/forks/versions/GossipForkSubscriptionsDeneb.java
@@ -17,14 +17,12 @@ import org.hyperledger.besu.plugin.services.MetricsSystem;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.networking.eth2.P2PConfig;
 import tech.pegasys.teku.networking.eth2.gossip.BlobSidecarGossipManager;
-import tech.pegasys.teku.networking.eth2.gossip.BlockAndBlobsSidecarGossipManager;
 import tech.pegasys.teku.networking.eth2.gossip.encoding.GossipEncoding;
 import tech.pegasys.teku.networking.eth2.gossip.topics.OperationProcessor;
 import tech.pegasys.teku.networking.p2p.discovery.DiscoveryNetwork;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.datastructures.attestation.ValidateableAttestation;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.deneb.SignedBeaconBlockAndBlobsSidecar;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.deneb.SignedBlobSidecar;
 import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
@@ -39,9 +37,7 @@ import tech.pegasys.teku.storage.client.RecentChainData;
 public class GossipForkSubscriptionsDeneb extends GossipForkSubscriptionsCapella {
 
   private final OperationProcessor<SignedBlobSidecar> blobSidecarProcessor;
-  private final OperationProcessor<SignedBeaconBlockAndBlobsSidecar> blockAndBlobsProcessor;
 
-  private BlockAndBlobsSidecarGossipManager blockAndBlobsSidecarGossipManager;
   private BlobSidecarGossipManager blobSidecarGossipManager;
 
   public GossipForkSubscriptionsDeneb(
@@ -55,7 +51,6 @@ public class GossipForkSubscriptionsDeneb extends GossipForkSubscriptionsCapella
       final GossipEncoding gossipEncoding,
       final OperationProcessor<SignedBeaconBlock> blockProcessor,
       final OperationProcessor<SignedBlobSidecar> blobSidecarProcessor,
-      final OperationProcessor<SignedBeaconBlockAndBlobsSidecar> blockAndBlobsProcessor,
       final OperationProcessor<ValidateableAttestation> attestationProcessor,
       final OperationProcessor<ValidateableAttestation> aggregateProcessor,
       final OperationProcessor<AttesterSlashing> attesterSlashingProcessor,
@@ -85,33 +80,13 @@ public class GossipForkSubscriptionsDeneb extends GossipForkSubscriptionsCapella
         signedContributionAndProofOperationProcessor,
         syncCommitteeMessageOperationProcessor,
         signedBlsToExecutionChangeOperationProcessor);
-
     this.blobSidecarProcessor = blobSidecarProcessor;
-    this.blockAndBlobsProcessor = blockAndBlobsProcessor;
   }
 
   @Override
   protected void addGossipManagers(final ForkInfo forkInfo) {
-    // TODO: change this to super.addGossipManagers(forkInfo) when the coupled block and blobs
-    // sidecar gossip manager is removed
-
-    // Phase0 without BlockGossipManager
-    addAttestationGossipManager(forkInfo);
-    addAggregateGossipManager(forkInfo);
-    addVoluntaryExitGossipManager(forkInfo);
-    addProposerSlashingGossipManager(forkInfo);
-    addAttesterSlashingGossipManager(forkInfo);
-
-    // Altair
-    addSignedContributionAndProofGossipManager(forkInfo);
-    addSyncCommitteeMessageGossipManager(forkInfo);
-
-    // Capella
-    addSignedBlsToExecutionChangeGossipManager(forkInfo);
-
-    // Deneb
+    super.addGossipManagers(forkInfo);
     addBlobSidecarGossipManager(forkInfo);
-    addBlockAndBlobsSidecarGossipManager(forkInfo);
   }
 
   void addBlobSidecarGossipManager(final ForkInfo forkInfo) {
@@ -128,28 +103,8 @@ public class GossipForkSubscriptionsDeneb extends GossipForkSubscriptionsCapella
     addGossipManager(blobSidecarGossipManager);
   }
 
-  void addBlockAndBlobsSidecarGossipManager(final ForkInfo forkInfo) {
-    blockAndBlobsSidecarGossipManager =
-        new BlockAndBlobsSidecarGossipManager(
-            recentChainData,
-            spec,
-            asyncRunner,
-            discoveryNetwork,
-            gossipEncoding,
-            forkInfo,
-            blockAndBlobsProcessor,
-            getMessageMaxSize());
-    addGossipManager(blockAndBlobsSidecarGossipManager);
-  }
-
   @Override
   public void publishBlobSidecar(final SignedBlobSidecar blobSidecar) {
     blobSidecarGossipManager.publishBlobSidecar(blobSidecar);
-  }
-
-  @Override
-  public void publishBlockAndBlobsSidecar(
-      final SignedBeaconBlockAndBlobsSidecar blockAndBlobsSidecar) {
-    blockAndBlobsSidecarGossipManager.publishBlockAndBlobsSidecar(blockAndBlobsSidecar);
   }
 }