+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/peers/SyncSource.java
@@ -30,8 +30,13 @@ public interface SyncSource {
   SafeFuture<Void> requestBlocksByRange(
       UInt64 startSlot, UInt64 count, RpcResponseListener<SignedBeaconBlock> listener);
 
-  SafeFuture<Void> requestBlobsSidecarsByRange(
-      UInt64 startSlot, UInt64 count, RpcResponseListener<BlobsSidecar> listener);
+  // TODO: remove when blobs decoupling sync is implemented
+  @Deprecated
+  @SuppressWarnings("unused")
+  default SafeFuture<Void> requestBlobsSidecarsByRange(
+      UInt64 startSlot, UInt64 count, RpcResponseListener<BlobsSidecar> listener) {
+    return SafeFuture.failedFuture(new UnsupportedOperationException());
+  }
 
   SafeFuture<Void> requestBlobSidecarsByRange(
       UInt64 startSlot, UInt64 count, RpcResponseListener<BlobSidecar> listener);