+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/forks/versions/GossipForkSubscriptionsDenebTest.java
@@ -86,7 +86,6 @@ public class GossipForkSubscriptionsDenebTest {
         noopOperationProcessor,
         noopOperationProcessor,
         noopOperationProcessor,
-        noopOperationProcessor,
         noopOperationProcessor);
   }
 }